mfd: twl6030-gpadc add driver for GPADC block
This patch adds support for the General Purpose A/D Convertor in
twl6030/twl6032 chips.

This driver is based on the twl4030-madc driver written by Mikko Ylinen
and Filipe Balbi, and currently has a 'misc' device interface. It needs
to be moved to drivers/hwmon as per comments from the community on the
twl4030-madc driver.

The GPADC blocks in the two chips have some small differences, twl6030
can measure X number of 10 bit channels. twl6032 can only measure one
channel at a time but has 12 bits of resolution. The driver hides the
differences in number of channel handling but not resolution from the
caller.

Signed-off-by: Nishant Kamat <nskamat@ti.com>
Signed-off-by: Balaji T K <balajitk@ti.com>
Signed-off-by: Oleksandr Kozaruk <oleksandr.kozaruk@ti.com>
Signed-off-by: Girish S G <girishsg@ti.com>
Signed-off-by: Ambresh K <ambresh@ti.com>
Signed-off-by: Graeme Gregory <gg@slimlogic.co.uk>
Signed-off-by: Volodymyr Riazantsev <v.riazantsev@ti.com>
Signed-off-by: Sergii Postulga <x0153364@ti.com>
Signed-off-by: Eugen Mandrenko <ievgen.mandrenko@ti.com>
Signed-off-by: Oleksandr Dmytryshyn <oleksandr.dmytryshyn@ti.com>
Signed-off-by: x0169938 <vasyl.yushchyshen@ti.com>