mfd: palmas add new varient with battery charger
Palmas Charger is a new member of the Palmas family with a battery
charger, more GPIO, more LEDs, more LDOs and some changes to some
of the SMPS type regulators.

Most signifcantly for MFD these new features add 2 new banks of IRQs so
we now need 3 regmap IRQs to handle them. So split the IRQ handling off
to a seperate file as it is now more complex.

This commit also add the new structure definitions for the changed/new
children and the new registers and defines needed.

Signed-off-by: Graeme Gregory <gg@slimlogic.co.uk>